requirements.txt
elasticsearch
fastapi==0.115.13
fastmcp==2.9.0
httpx==0.28.1
openai==1.91.0
python-dotenv==1.1.1
Requests==2.32.4
sseclient==0.0.27
uvicorn==0.34.3
